SDK/RDS/requirements.txt
jinja2<3.1.0
jinjasql
psycopg2-binary
pandas
dateutil